backend/requirements.txt
Flask==3.0.3
Flask-Cors== 5.0.0
python-dotenv==1.0.1
networkx== 3.4.2
numpy== 2.1.3
matplotlib== 3.9.2
